refactor(mcp): move work handlers behind the tool context
