+++ crates/tracedecay-mcp/src/handlers/mod.rs
@@ -19,6 +19,7 @@ mod multi_root;
 pub mod redundancy;
 mod retained_response;
 pub mod support;
+pub mod work;
 
 pub use bounded_search::run_bounded_search;
 pub use multi_root::handle_multi_root;
@@ -31,3 +32,4 @@ pub use support::{
     take_internal_context_memory_analytics, text_tool_result, tool_json, tool_json_with_md,
     unique_file_paths,
 };
+pub use work::handle_work;
