+++ crates/tracedecay/src/mcp/tools/handlers/mod.rs
@@ -121,7 +121,15 @@ mod tool_definition_tests;
     clippy::uninlined_format_args
 )]
 mod verified_graph_query_authority_tests;
-mod work;
+#[cfg(test)]
+#[allow(
+    clippy::unwrap_used,
+    clippy::expect_used,
+    clippy::await_holding_lock,
+    clippy::redundant_closure_for_method_calls,
+    clippy::uninlined_format_args
+)]
+mod work_dispatch_tests;
 pub mod workflow;
 mod workflow_family;
 
@@ -159,13 +167,14 @@ use retained_catalog::dispatch_profile_retained_application_tool;
 use retained_catalog::retained_mcp_composition;
 pub(crate) use tool_call_support::INTERNAL_DAEMON_TOOL_NAMES;
 use tool_call_support::{boxed_send, rejected_tool_project_selector_present};
+use tracedecay_api::WorkHttpRequest;
 use tracedecay_contracts::ProjectRegistryReadPort;
+use tracedecay_daemon_protocol::DaemonInvocationExecutor;
 use tracedecay_daemon_service::application_surface::resolve_catalog_tool_binding;
 use tracedecay_domain::errors::{Result, TraceDecayError};
 use tracedecay_global_db::RegisteredGlobalDbLeaseV1;
 use tracedecay_mcp::ToolResult;
-use tracedecay_mcp::handle_multi_root;
-use work::handle_work;
+use tracedecay_mcp::{handle_multi_root, handle_work};
 use workflow_family::handle_workflow;
 
 /// Dispatches a tool call to the appropriate handler.
@@ -568,7 +577,9 @@ pub fn handle_tool_call_with_registry_options<'a>(
             return boxed_send(handle_work(
                 tool_name,
                 args,
-                options.application_invocation_executor,
+                options.application_invocation_executor.map(|executor| {
+                    move |request| invoke_admitted_work_operation(executor, request)
+                }),
                 options.application_request_id,
                 options.application_deadline,
                 options.application_cancellation,
@@ -790,6 +801,32 @@ fn seated_generation_age_label(sealed_at: tracedecay_domain::UtcMicros) -> Strin
     } else {
         format!("{}d", seconds / 86_400)
     }
+}
+
+/// Reads the canonical Work HTTP envelope the daemon owner already produced.
+async fn invoke_admitted_work_operation(
+    executor: &dyn DaemonInvocationExecutor,
+    request: WorkHttpRequest,
+) -> Result<Value> {
+    let response =
+        tracedecay_daemon_service::application_surface::invoke_work_operation(executor, request)
+            .await;
+    let body = axum::body::to_bytes(response.into_body(), usize::MAX)
+        .await
+        .map_err(|error| {
+            TraceDecayError::project_route(
+                "work.response_unavailable",
+                true,
+                format!("The Work application response could not be read: {error}"),
+            )
+        })?;
+    serde_json::from_slice(&body).map_err(|error| {
+        TraceDecayError::project_route(
+            "work.response_invalid",
+            true,
+            format!("The Work application response was not valid JSON: {error}"),
+        )
+    })
 }
 
 /// The single rejection every dispatch group returns for a name it does not own.
